Completed mouseover color change. Pixels stay dark
Code updated to include a mouseover event that is applied to each pixel
so that each pixel turns colors. I have also added a mousedown event
that is mostly useless for the moment. I will be committing this change
and moving on to try and rework this event for better performance. The
mouseover event tends to skip a lot of cells since the calculation
is currently so taxing.