Branch Button now generally working on button
User is now prompted on button click instead of page load for the
resolution number. In addition, the program logic is now simplified with
some basic start programming and the button load, and the button now
launches the remaining program successively.

Further edits to be made to the program. The button does not clear the
previous field from view and merely appends more etchasketch fields to
the page. A new function to create these fields will be added in next
update.